requirements.txt
cryptography
pycryptodome
PyQt5
binwalk
Pillow
pyzbar
py7zr
rarfile
requests
